+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+# Head
+
+-   Refactor: simplify code.
+
 # 3.0.0 - 2017-06-20
 
 -   Chore: support `webpack` v3.